remove code duplication in inspector
