+++ mscore/inspector/inspectorTextLine.h
@@ -14,9 +14,7 @@
 #define __INSPECTOR_TEXTLINE_H__
 
 #include "inspector.h"
-#include "inspectorBase.h"
-#include "ui_inspector_textline.h"
-#include "ui_inspector_line.h"
+#include "inspectorTextLineBase.h"
 
 namespace Ms {
 
@@ -24,19 +22,12 @@ namespace Ms {
 //   InspectorTextLine
 //---------------------------------------------------------
 
-class InspectorTextLine : public InspectorElementBase {
+class InspectorTextLine : public InspectorTextLineBase {
       Q_OBJECT
-
-      Ui::InspectorLine l;
-      Ui::InspectorTextLine tl;
 
    public:
       InspectorTextLine(QWidget* parent);
-      virtual void setElement() override;
       };
-
-extern void populateHookType(QComboBox* b);
-extern void populateTextPlace(QComboBox* b);
 
 } // namespace Ms
 #endif
