+++ mscore/inspector/inspectorTextLineBase.cpp
@@ -19,6 +19,32 @@
 namespace Ms {
 
 //---------------------------------------------------------
+//   populateHookType
+//---------------------------------------------------------
+
+void populateHookType(QComboBox* b)
+      {
+      b->clear();
+      b->addItem(b->QObject::tr("None"), int(HookType::NONE));
+      b->addItem(b->QObject::tr("90°"), int(HookType::HOOK_90));
+      b->addItem(b->QObject::tr("45°"), int(HookType::HOOK_45));
+      b->addItem(b->QObject::tr("90° centered"), int(HookType::HOOK_90T));
+      }
+
+//---------------------------------------------------------
+//   populateTextPlace
+//---------------------------------------------------------
+
+void populateTextPlace(QComboBox* b)
+      {
+      b->clear();
+      b->addItem(b->QObject::tr("Auto"),  int(PlaceText::AUTO));
+      b->addItem(b->QObject::tr("Above"), int(PlaceText::ABOVE));
+      b->addItem(b->QObject::tr("Below"), int(PlaceText::BELOW));
+      b->addItem(b->QObject::tr("Left"),  int(PlaceText::LEFT));
+      }
+
+//---------------------------------------------------------
 //   InspectorTextLineBase
 //---------------------------------------------------------
